requirements.txt
langchain-google-genai
langgraph
langchain-chroma
pypdf
pymupdf
pdfplumber
streamlit
python-dotenv
chromadb
pandas